chore: move component to components folder
