+++ src/index.tsx
@@ -1,5 +1,5 @@
 import {requireNativeComponent, UIManager, Platform, ViewStyle} from 'react-native'
-import Button from './Button'
+import Button from './components/Button'
 
 const LINKING_ERROR = `The package 'rn-base-component' doesn't seem to be linked. Make sure: \n\n ${Platform.select(
   {ios: "- You have run 'pod install'\n", default: ''},
